refactor: Follow 관련 기능 리팩토링
- DB에 존재하는지 혹은 존재하지 않는지 등의 체크 로직 추가

+++ src/main/java/com/example/naejango/domain/follow/application/FollowService.java
@@ -39,6 +39,12 @@ public class FollowService {
 
     /** 창고 팔로우 등록 */
     public void addFollow(Long userId, Long storageId){
+        // 이미 팔로우 등록 되어있는지 체크, DB에 존재하면 true return
+        boolean existCheck = followRepository.existsByUserIdAndStorageId(userId, storageId);
+        if (existCheck) {
+            throw new CustomException(ErrorCode.FOLLOW_ALREADY_EXIST);
+        }
+
         Storage storage = storageRepository.findById(storageId)
                 .orElseThrow(() -> new CustomException(ErrorCode.STORAGE_NOT_FOUND));
         User user = userRepository.findById(userId)
@@ -57,6 +63,9 @@ public class FollowService {
     /** 창고 팔로우 해제 */
     public void deleteFollow(Long userId, Long storageId){
         Follow follow = followRepository.findByUserIdAndStorageId(userId, storageId);
+        if (follow == null) {
+            throw new CustomException(ErrorCode.FOLLOW_NOT_FOUND);
+        }
 
         followRepository.delete(follow);
     }
